+++ src/Toolkit/tests/Kit/Lint/Checker/DocCheckerTest.php
@@ -18,7 +18,6 @@ use Symfony\UX\Toolkit\Kit\Lint\Checker\DocChecker;
 use Symfony\UX\Toolkit\Kit\Lint\LintSeverity;
 use Symfony\UX\Toolkit\Recipe\Recipe;
 use Symfony\UX\Toolkit\Recipe\RecipeManifest;
-use Symfony\UX\Toolkit\Recipe\RecipeType;
 
 final class DocCheckerTest extends TestCase
 {
@@ -38,6 +37,6 @@ final class DocCheckerTest extends TestCase
 
     private function recipeManifest(string $name): RecipeManifest
     {
-        return new RecipeManifest(RecipeType::Component, $name, []);
+        return new RecipeManifest($name, []);
     }
 }